+++ src/main/java/techcourse/myblog/domain/User.java
@@ -20,7 +20,7 @@ public class User {
     private String email;
     private String password;
 
-    public User() {
+    private User() {
     }
 
     public User(String name, String email, String password) {
