+++ src/test/java/techcourse/myblog/domain/UserTest.java
@@ -1,11 +1,8 @@
 package techcourse.myblog.domain;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class UserTest {
     @Test
